Replace PApplet by PShape
Otherwise we can not specify in which PGraphics we want to create the shape. Sometimes it's important to be able to choose. I noticed that if I don't specify the right PGraphics nothing is rendered.

Also replace PApplet.LINES by PConstants.LINES which I think it's the recommended approach for using the Processing constants.